+++ src/pages/index.astro
@@ -1,5 +1,5 @@
 ---
-import Menu from "ui/Menu";
+import Menu from "components/Menu";
 import Layout from "../layouts/Layout.astro";
 ---
