+++ tsconfig.json
@@ -27,7 +27,7 @@
     "paths": {
       "context": ["common/context/index.tsx"],
       "state":[ "common/state.ts"],
-      "ui": ["ui/index.tsx"],
+      "components": ["components/index.tsx"],
       "utils": ["common/utils.ts"],
       "styles":[ "styles/*"]
     }
